third_party/blink/web_tests/fast/filesystem/file-writer-abort-continue-expected.txt

Test that FileWriter can continue immediately after an abort.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS 1100000 is blobSize
starting test
PASS Calling write.
PASS Calling abort
PASS Saw abort
PASS writer.length is 0
PASS Saw writeend 0.
PASS Calling write.
PASS writer.length is 1100000
PASS Saw writeend 1.
PASS Calling truncate.
PASS writer.length is 0
PASS Saw writeend 2.
PASS Calling write.
PASS Calling abort at the end of the write
PASS Saw abort
PASS writer.length is 1100000
PASS Saw writeend 3.
PASS Calling write.
PASS writer.length is 2200000
PASS Saw writeend 4.
PASS Calling truncate.
PASS writer.length is 0
PASS Saw writeend 5.
PASS Calling write.
PASS Calling abort
PASS Saw abort
PASS writer.length is 0
PASS Saw writeend 6.
PASS Calling truncate.
PASS writer.length is 7
PASS Saw writeend 7.
PASS All tests complete.
PASS successfullyParsed is true

TEST COMPLETE
